Homework 15 - Spring Boot: refactor, logic fixes - 2

+++ src/test/java/org/itrex/repositories/impl/HibernateRoleRepoTest.java
@@ -56,7 +56,7 @@ public class HibernateRoleRepoTest extends TestBaseHibernate {
     @DisplayName("getRoleById with valid data - should return a Role with given id")
     public void getRoleById1() {
         // given
-        long roleId = 1L;
+        Integer roleId = 1;
 
         // when
         Role role = repo.getRoleById(roleId);
@@ -70,7 +70,7 @@ public class HibernateRoleRepoTest extends TestBaseHibernate {
     @DisplayName("getRoleById with invalid data - should throw DatabaseEntryNotFoundException")
     public void getRoleById2() {
         // given
-        long roleId = 150L;
+        Integer roleId = 150;
 
         // when & then
         assertThrows(DatabaseEntryNotFoundException.class, () -> repo.getRoleById(roleId));
